docs: 필요없는 코드 정리 및 수정
