+++ src/Components/CompanyRecruitment/Table/index.tsx
@@ -5,7 +5,6 @@ import { Pagination } from '../../../Utils/Pagination';
 import { useChangeCompanyStatus, useChangeContractCompany } from '../../../Apis/Companies';
 import { dataType } from '../../../Apis/Companies/request';
 import { CompanyRecruitmentResponse } from '../../../Apis/Companies/response';
-import { getValueByKey } from '../../../Hooks/useGetPropertyValueAndKey';
 import { companyType } from '../../../Utils/Translation';
 
 interface PropsType {
@@ -75,7 +74,7 @@ export function CompanyRecruitmentTable({ companyRecruitment, refetchCompanyRecr
 				<_.ContentText>{companie.business_area}</_.ContentText>, // 사업분야
 				<_.ContentText>{companie.workers_count}</_.ContentText>, // 근로자수
 				<_.ContentText>{companie.take}</_.ContentText>, // 매출액
-				<_.ContentText status={companie.company_type === 'PARTICIPATING'}>{getValueByKey(companyType, companie.company_type)}기업</_.ContentText>, // 기업구분
+				<_.ContentText status={companie.company_type === 'PARTICIPATING'}>{companyType[companie.company_type]}기업</_.ContentText>, // 기업구분
 				<_.ContentText>{companie.convention && 'Y'}</_.ContentText>, // 협약여부
 				<_.ContentText>{companie.personal_contact && 'Y'}</_.ContentText>, // 개인컨택
 				<_.ContentText>{companie.recent_recruit_year ? `${companie.recent_recruit_year}년` : '없음'}</_.ContentText>, //최근의뢰년도
