+++ src/Components/Modal/ChangeStatusModal/index.tsx
@@ -1,4 +1,3 @@
-import { getValueByKey } from '../../../Hooks/useGetPropertyValueAndKey';
 import { applicationStatus } from '../../../Utils/Translation';
 import * as _ from '../style';
 
@@ -13,8 +12,8 @@ export function ChangeStatusModal({ clickedData, clickStudentName, statusName }:
 		<>
 			<_.ModalTitleText>
 				{clickedData.length - 3 > 0
-					? `${clickStudentName.join(', ')} 외 ${clickedData.length - 3}명의 지원서를 ${getValueByKey(applicationStatus, statusName)}(으)로 바꾸시겠습니까?`
-					: `${clickStudentName.join(', ')}의 지원서를 ${getValueByKey(applicationStatus, statusName)}(으)로 바꾸시겠습니까?`}
+					? `${clickStudentName.join(', ')} 외 ${clickedData.length - 3}명의 지원서를 ${applicationStatus[statusName]}(으)로 바꾸시겠습니까?`
+					: `${clickStudentName.join(', ')}의 지원서를 ${applicationStatus[statusName]}(으)로 바꾸시겠습니까?`}
 			</_.ModalTitleText>
 			<_.ModalContentText>삭제하면 되돌릴 수 없습니다.</_.ModalContentText>
 		</>
