+++ src/Utils/Translation.ts
@@ -1,11 +1,11 @@
-export const companyType = {
+export const companyType: { [key: string]: string } = {
 	LEAD: '선도',
 	PARTICIPATING: '참여',
 	CONTRACTING: '협약',
 	DEFAULT: '기본',
 };
 
-export const applicationStatus = {
+export const applicationStatus: { [key: string]: string } = {
 	REQUESTED: '승인요청',
 	APPROVED: '승인',
 	FAILED: '불합격',
@@ -13,7 +13,7 @@ export const applicationStatus = {
 	REJECTED: '반려',
 };
 
-export const companyStatus = {
+export const companyStatus: { [key: string]: string } = {
 	전체: '',
 	모집전: 'READY',
 	모집중: 'RECRUITING',
@@ -21,7 +21,7 @@ export const companyStatus = {
 	접수요청: 'REQUESTED',
 };
 
-export const applicationStatusTextColor = {
+export const applicationStatusTextColor: { [key: string]: string } = {
 	REQUESTED: '#F1C40F',
 	APPROVED: '#2ECC71',
 	FAILED: '#E74C3C',
